+++ lib/flow_avatar.dart
@@ -6,5 +6,6 @@ library;
 
 export 'src/avatar_model.dart' show FlowAvatarModel, FlowAvatarSpot;
 export 'src/avatar_seed.dart' show flowAvatarSeed;
-export 'src/avatar_state.dart' show FlowAvatarShape, FlowAvatarState;
+export 'src/avatar_state.dart'
+    show FlowAvatarShape, FlowAvatarState, flowAvatarStateSpeed;
 export 'src/flow_avatar.dart' show FlowAvatar;